Added Synthetic_Female_Phonetics_0001 Voice (#1104)
* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Update Bully.txt

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Update Bully.txt

Improved the prompt.

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Update Bully.txt

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Update Bully.txt

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Delete Synthetic_Duke_Nukem.wav

* Add files via upload

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Listen to voice first by generating it in AllTalk before listening to the wav file. I am using a technique called Phonetics.  This voice is trained on 4 different synthetic voices that I created. The output with AllTalk is so good creating a single new realistic voice from 4 different synthetic voices without using spoken words for the training. You do not need to train AllTalk voices on voice lines it can also be trained on with Phonetics. The wav might sound funny for humans but it is perfect for the Alltalk A.I

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

* Add files via upload

Added Synthetic_Female_Phonetics_0001 Voice
This is a voice file for AllTalk.

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>

---------

Signed-off-by: CRCODE22 <88407346+CRCODE22@users.noreply.github.com>